SIP refactoring
