+++ src/main/java/interfaces/CarAccessInterface.java
@@ -12,10 +12,10 @@ public interface CarAccessInterface {
      * Iterare through the car dataset and return all cars
      * @return a list of Car objects from the dataset by ID
      */
-    public ResultSet getCar(int id);
+    ResultSet getCar(int id);
 
     /**
      * @return the Car entity representation of the car.
      */
-    public ResultSet getAllCars();
+    ResultSet getAllCars();
 }
