--- src/main/java/interfaces/CarOutputAdapter.java
@@ -1,8 +1,0 @@
-package interfaces;
-
-import entities.Car;
-
-public interface CarOutputAdapter {
-
-    String request(Car car);
-}
