+++ src/main/java/interfaces/SensoJsonInterface.java
@@ -8,5 +8,5 @@ public interface SensoJsonInterface {
     /**
      * @return a JSON representation of the data needed for the Senso API body for the Loan API call.
      */
-    public String sensoReadyData();
+    String sensoReadyData();
 }
